Improve http4s url encoding tests.
