+++ http4s/client/js/src/test/scala/endpoints4s/http4s/client/Http4sClientUrlEncodingTestJS.scala
@@ -1,0 +1,18 @@
+package endpoints4s.http4s.client
+
+import cats.effect.IO
+import org.http4s.Uri
+import org.http4s.dom.FetchClientBuilder
+
+class Http4sClientUrlEncodingTestJS
+    extends Http4sClientUrlEncodingTest[IO, TestJsonSchemaClient[IO]] {
+
+  val client = new TestJsonSchemaClient[IO](
+    Uri.Authority(
+      host = Uri.RegName("localhost"),
+      port = Some(8080)
+    ),
+    Uri.Scheme.http,
+    FetchClientBuilder[IO].create
+  )
+}
